Fixing errors due to import being async

+++ src/file-io/todo-yaml-ensure.js
@@ -3,30 +3,37 @@ import defaultTodoFile from "../default-todo.js";
 import inquirer from "inquirer";
 import { TODO_FILE_PATH } from "../constants.js";
 
-let fileExists = true;
-try {
-    await access(TODO_FILE_PATH);
-} catch {
-    fileExists = false;
-}
-if (!fileExists) {
+let fileExists = false;
+export async function ensureTodoYamlFile() {
+    if (fileExists) {
+        return;
+    }
     try {
-        const answers = await inquirer
-            .prompt([
-                {
-                    name: "Confirm",
-                    type: "confirm",
-                    message: "todo.yaml not found. Create a todo.yaml in this directory?",
-                    default: false
-                }
-            ]);
-        if (answers.Confirm) {
-            await writeFile("todo.yaml", defaultTodoFile);
-        } else {
-            console.log("Please go to the directory you want your todo.yaml, and try again.");
-            process.exit();
+        await access(TODO_FILE_PATH);
+        fileExists = true;
+    } catch {
+        fileExists = false;
+    }
+    if (!fileExists) {
+        try {
+            const answers = await inquirer
+                .prompt([
+                    {
+                        name: "Confirm",
+                        type: "confirm",
+                        message: "todo.yaml not found. Create a todo.yaml in this directory?",
+                        default: false
+                    }
+                ]);
+            if (answers.Confirm) {
+                await writeFile("todo.yaml", defaultTodoFile);
+            } else {
+                console.log("Please go to the directory you want your todo.yaml, and try again.");
+                process.exit();
+            }
+        } catch (err) {
+            console.log(`Whoops. Couldn't create todo.yaml: ${err}`);
         }
-    } catch (err) {
-        console.log(`Whoops. Couldn't create todo.yaml: ${err}`);
     }
 }
+
